update profile latest

+++ database/migrations/2024_12_30_020725_add_role_id_to_users_table.php
@@ -11,10 +11,12 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('users', function (Blueprint $table) {
+    Schema::table('users', function (Blueprint $table) {
+        if (!Schema::hasColumn('users', 'role_id')) {
             $table->unsignedBigInteger('role_id')->nullable()->after('password'); // Add role_id
             $table->foreign('role_id')->references('id')->on('roles')->onDelete('set null'); // Foreign key constraint
-        });
+        }
+    });
     }
 
     /**
